+++ source/extensions/lightspeed.trex.contexts/docs/CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.0.2] - 2024-04-12
+### Added
+- Global events for context and page changes
+
 ## [1.0.1]
 ### Changed
 - Set Apache 2 license headers