+++ source/extensions/lightspeed.trex.properties_pane.shared.mod_setup.widget/docs/CHANGELOG.md
@@ -1,6 +1,10 @@
 ﻿# Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.4.0] - 2024-04-12
+### Changed
+- Project Wizard Window now references global instance
+
 ## [1.3.2]
 ### Changed
 - Set Apache 2 license headers